+++ src/core/hw/tegra_x1/cpu/cpu.hpp
@@ -25,7 +25,7 @@ class ICpu {
     virtual IThread* CreateThread(WallClock& wall_clock, IMmu* mmu,
                                   const ThreadCallbacks& callbacks,
                                   IMemory* tls_mem, vaddr_t tls_mem_base) = 0;
-    virtual IMemory* AllocateMemory(usize size) = 0;
+    virtual IMemory* AllocateMemory(u64 size) = 0;
 
   protected:
     CpuFeatures features;
