+++ src/core/hw/tegra_x1/cpu/dynarmic/cpu.hpp
@@ -24,7 +24,7 @@ class Cpu : public ICpu {
     IThread* CreateThread(WallClock& wall_clock, IMmu* mmu,
                           const ThreadCallbacks& callbacks, IMemory* tls_mem,
                           vaddr_t tls_mem_base) override;
-    IMemory* AllocateMemory(usize size) override;
+    IMemory* AllocateMemory(u64 size) override;
 };
 
 } // namespace hydra::hw::tegra_x1::cpu::dynarmic
